pyproject.toml
[project]
name = "pushorpay"
version = "0.1.0"
description = ""
authors = [
    {name = "dsmai",email = "son.dmai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.16.0,<3.0.0)",
    "postgrest (>=1.1.1,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "pushorpay", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.3"
mypy = "^1.17.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
